@@ -5,7 +5,8 @@ module CreateSession
   def jwt_session_create(user_id, user_email)
     user = User.find_by(id: user_id)
     session = user.sessions.build
+    return unless user && session.save
 
-    JsonWebToken.encode({ user_id: user_id, user_email: user_email, user_role: user.role, token: session.token }) if user && session.save
+    JsonWebToken.encode({ user_id: user_id, user_email: user_email, user_role: user.role, token: session.token })
   end
 end
