@@ -21,36 +21,9 @@ export class BugbountyStatComponent implements OnInit {
 
   loadSyncInti=false
   loadSyncYwh=false
-  scopeYWH = [
-    {
-      "id":"",
-      "title":"",
-      "severity":"",
-      "reward":"",
-      "collab":"",
-      "status":"",
-      "report_date":""
-    }
-  ]
+  scopeYWH = this.scopeTemplate()
 
-  statYWH = {
-    earnedEuro:0,
-    collab_number:0,
-    total_rapports:0,
-    average_per_rapport:0.0,
-    rapport_severity:{
-      "C":0,
-      "H":0,
-      "M":0,
-      "L":0
-    },
-    report_by_month:{
-    },
-    report_by_status:{
-    },
-    earn_by_month:{
-    }
-  };
+  statYWH = this.statTemplate()
   optionsYWHPie:any ;
   optionsYWHPieReportStatus: any;
   optionsYWHBarre: any = {};
@@ -63,35 +36,8 @@ optionsINTIPieReportStatus: any;
 optionsINTIBarre: any = {};
 optionsINTIEarnedByMonth: any = {};
 
-statINTI = {
-  earnedEuro:0,
-  collab_number:0,
-  total_rapports:0,
-  average_per_rapport:0.0,
-  rapport_severity:{
-    "C":0,
-    "H":0,
-    "M":0,
-    "L":0
-  },
-  report_by_month:{
-  },
-  report_by_status:{
-  },
-  earn_by_month:{
-  }
-};
-  scopeINTI = [
-    {
-      "id":"",
-      "title":"",
-      "severity":"",
-      "reward":"",
-      "collab":"",
-      "status":"",
-      "report_date":""
-    }
-  ]
+statINTI = this.statTemplate()
+scopeINTI = this.scopeTemplate()
 
 
   loading=true
@@ -113,18 +59,15 @@ statINTI = {
     this.bugbountyPlatform.getPlatform().subscribe( (result) => {
       this.yeswehackExist = false;
       this.intigritiExist = false;
-      let one = false
       result.data.forEach( (element) => {
        // console.log(this.yeswehackExist)
         if(element.name == 'yeswehack'){
           this.yeswehackExist=true
           this.getStatsPlatform('yeswehack')
-          one=true
         } 
         if(element.name == 'intigriti'){
           this.intigritiExist = true
           this.getStatsPlatform('intigriti')
-          one=true
         } 
       });     
       this.loading = false;
@@ -134,54 +77,92 @@ statINTI = {
     })
 
   }
+  statTemplate(){
+    return {
+      earnedEuro:0,
+      collab_number:0,
+      total_rapports:0,
+      average_per_rapport:0.0,
+      rapport_severity:{
+        "C":0,
+        "H":0,
+        "M":0,
+        "L":0
+      },
+      report_by_month:{
+      },
+      report_by_status:{
+      },
+      earn_by_month:{
+      }
+    };
+  }
+  scopeTemplate(){
+    return [
+      {
+        "id":"",
+        "title":"",
+        "severity":"",
+        "reward":"",
+        "collab":"",
+        "status":"",
+        "report_date":""
+      }
+    ]
+  }
+  countSeverity(severity:string,rapport_severity:any){
+    if(severity=="Low")this.statYWH.rapport_severity.L++
+    if(severity=="Medium")rapport_severity.M++
+    if(severity=="High")rapport_severity.H++
+    if(severity=="Critical")rapport_severity.C++
+    return rapport_severity
+  }
   computeYWHStat() {
-    var temp=Array();
-    var tempStatus=Array()
-    var tempEarnByMonth=Array()
-    this.loadingYWHGlobal=true
-    _.each(this.scopeYWH, (element) => {
-      //total earned
-      this.statYWH.earnedEuro +=  parseInt(element.reward)
-      // collab repport
-      if(element.collab){
-        this.statYWH.collab_number++
-      }
-      // order by severety
-      if(element.severity=="Low")this.statYWH.rapport_severity.L++
-      if(element.severity=="Medium")this.statYWH.rapport_severity.M++
-      if(element.severity=="High")this.statYWH.rapport_severity.H++
-      if(element.severity=="Critical")this.statYWH.rapport_severity.C++
-      //nombre de rapport par mois
-      let date = new Date(element.report_date)
-      let dateMonth = `${date.getMonth()+1}`.padStart(2, '0');
-      let date2 = parseInt(`${date.getFullYear()}${dateMonth}`)
-      if(_.isUndefined(temp[date2])){
-        temp[date2]=0
-        tempEarnByMonth[date2]=0
-      }
-      tempEarnByMonth[date2]+=element.reward
-      temp[date2]++
-      //gain par mois
-
-      //status des rapports
-      if(_.isUndefined(tempStatus[element.status as keyof typeof tempStatus]))tempStatus[element.status as keyof typeof tempStatus]=0
-      tempStatus[element.status as keyof typeof tempStatus]++
-
-    })
-    
-    this.statYWH.report_by_month= temp
-    this.statYWH.report_by_status = tempStatus
-    this.statYWH.earn_by_month = tempEarnByMonth
-    this.statYWH.total_rapports=this.scopeYWH.length
-    this.statYWH.average_per_rapport = parseFloat((this.statYWH.earnedEuro /this.scopeYWH.length).toFixed(2))
+    this.statYWH = this.globalStat(this.scopeYWH)
     this.loadingYWHGlobal=false
-   // console.log(this.statYWH)
     this.optionsYWHPie = this.pieCriticity(this.statYWH.rapport_severity.L,this.statYWH.rapport_severity.M,this.statYWH.rapport_severity.H,this.statYWH.rapport_severity.C)
     this.initYWHBarre()
     this.initYWHPiReportStatus()
     this.initYWHEarnByMonth()
-    //console.log(this.statYWH)
-    
+  }
+  globalStat(scope:any):any{
+    let returnData:{report_by_month:any,earn_by_month:any,report_by_status:any,earnedEuro:number,collab_number:number,rapport_severity:any,average_per_rapport:number,total_repports:number}={
+      "report_by_month":{},
+      "earn_by_month":{},
+      "report_by_status":{},
+      "earnedEuro":0,
+      "collab_number":0,
+      "rapport_severity":{},
+      "average_per_rapport":0.0,
+      "total_repports":0
+    }
+    _.each(scope, (element) => {
+      console.log(returnData,element)
+      //total earned
+      returnData.earnedEuro +=  parseInt(element.reward)
+      // collab repport
+      if(element.collab){
+        returnData.collab_number++
+      }
+     returnData.rapport_severity=this.countSeverity(element.severity,this.statYWH.rapport_severity)
+      //nombre de rapport par mois
+      let date = new Date(element.report_date)
+      let dateMonth = `${date.getMonth()+1}`.padStart(2, '0');
+      let date2 = parseInt(`${date.getFullYear()}${dateMonth}`)
+      if(_.isUndefined(returnData.report_by_month[date2 as keyof typeof returnData.report_by_month])){
+        returnData.report_by_month[date2 as keyof typeof returnData.report_by_month]=0
+        returnData.earn_by_month[date2]=0
+      }
+      returnData.report_by_month[date2]+=element.reward
+      returnData.earn_by_month[date2]++
+      //status des rapports
+      if(_.isUndefined( returnData.report_by_status[element.status as keyof typeof returnData.report_by_status]))returnData.report_by_status[element.status as keyof typeof returnData.report_by_status]=0
+      returnData.report_by_status[element.status as keyof typeof returnData.report_by_status]++
+    })
+    returnData.total_repports=scope.length
+    returnData.average_per_rapport=parseFloat((returnData.earnedEuro /scope.length).toFixed(2))
+    console.log(returnData)
+    return returnData
   }
  
   syncYWH() {
@@ -196,39 +177,22 @@ statINTI = {
   }
 
   initYWHPiReportStatus(){
-    let data=Array()
-    Object.keys(this.statYWH.report_by_status).forEach( (element) => {
-      data.push({value:this.statYWH.report_by_status[element as keyof typeof this.statYWH.report_by_status],name:element})
-    })
-    //console.log(data)
+    let data   = this.initPie(this.statYWH)
     this.optionsYWHPieReportStatus = this.pieRepport(data)
   }
 
   initYWHBarre(){
-    let month=(new Date()).getMonth();
-    let year=((new Date()).getFullYear()-2);
-    let data=Array()
-    let label=Array()
-    for(var i=0;i<25;i++){
-      let padMonth = `${month}`.padStart(2, '0')
-      let key:string = `${year}${padMonth}`
-      let tempLabel:string = `${year}-${padMonth}`
-      if(Object.prototype.hasOwnProperty.call(this.statYWH.report_by_month,key)){
-       data.push(this.statYWH.report_by_month[key as keyof typeof this.statYWH.report_by_month])
-      } else {
-        data.push(0)
-      }
-      label.push(tempLabel)
-      if(month==12){
-        month=0
-        year++
-      }
-      month++
-    }
-    this.optionsYWHBarre = this.chartService.barreGraph(data,label,'Number of rapport','Number of report')
+   let data = this.initBarreReportByMonth(this.statYWH)
+   this.optionsYWHBarre = this.chartService.barreGraph(data[0],data[1],'Number of rapport','Number of report')
   }
-  
-  initYWHEarnByMonth(){
+  initPie(stat:any){
+    let data=Array()
+    Object.keys(stat.report_by_status).forEach( (element) => {
+      data.push({value:stat.report_by_status[element as keyof typeof stat.report_by_status],name:element})
+    })
+    return data
+  }
+  initBarreReportByMonth(stat:any){
     let month=(new Date()).getMonth();
     let year=((new Date()).getFullYear()-2);
     let data=Array()
@@ -237,8 +201,8 @@ statINTI = {
       let padMonth = `${month}`.padStart(2, '0')
       let key:string = `${year}${padMonth}`
       let tempLabel:string = `${year}-${padMonth}`
-      if(Object.prototype.hasOwnProperty.call(this.statYWH.earn_by_month,key)){
-       data.push(this.statYWH.earn_by_month[key as keyof typeof this.statYWH.earn_by_month])
+      if(Object.prototype.hasOwnProperty.call(stat.report_by_month,key)){
+       data.push(stat.report_by_month[key as keyof typeof stat.report_by_month])
       } else {
         data.push(0)
       }
@@ -249,7 +213,34 @@ statINTI = {
       }
       month++
     }
-    this.optionsYWHEarnedByMonth = this.chartService.barreGraph(data,label,"Earn by month",'value')
+    return [data,label]
+  }
+  initBarreEarnedByMonth(stat:any){
+    let month=(new Date()).getMonth();
+    let year=((new Date()).getFullYear()-2);
+    let data=Array()
+    let label=Array()
+    for(var i=0;i<25;i++){
+      let padMonth = `${month}`.padStart(2, '0')
+      let key:string = `${year}${padMonth}`
+      let tempLabel:string = `${year}-${padMonth}`
+      if(Object.prototype.hasOwnProperty.call(stat.earn_by_month,key)){
+       data.push(stat.earn_by_month[key as keyof typeof stat.earn_by_month])
+      } else {
+        data.push(0)
+      }
+      label.push(tempLabel)
+      if(month==12){
+        month=0
+        year++
+      }
+      month++
+    }
+    return [data,label]
+  }
+  initYWHEarnByMonth(){
+    let data =  this.initBarreEarnedByMonth(this.statYWH)
+    this.optionsYWHEarnedByMonth = this.chartService.barreGraph(data[0],data[1],"Earn by month",'value')
   }
   //---------------Intigriti part
 
@@ -268,54 +259,12 @@ statINTI = {
   }
 
   computeINTIStat() {
-      var temp=Array();
-      var tempStatus=Array()
-      var tempEarnByMonth=Array()
-      this.loadingINTIGlobal=true
-      console.log(this.scopeINTI)
-      _.each(this.scopeINTI, (element) => {
-        //total earned
-        this.statINTI.earnedEuro +=  parseInt(element.reward)
-        // collab repport
-        if(element.collab){
-          this.statINTI.collab_number++
-        }
-        // order by severety
-        if(element.severity=="Low")this.statINTI.rapport_severity.L++
-        if(element.severity=="Medium")this.statINTI.rapport_severity.M++
-        if(element.severity=="High")this.statINTI.rapport_severity.H++
-        if(element.severity=="Critical")this.statINTI.rapport_severity.C++
-        //nombre de rapport par mois
-        let date = new Date(element.report_date)
-        let dateMonth = `${date.getMonth()+1}`.padStart(2, '0');
-        let date2 = parseInt(`${date.getFullYear()}${dateMonth}`)
-        if(_.isUndefined(temp[date2])){
-          temp[date2]=0
-          tempEarnByMonth[date2]=0
-        }
-        tempEarnByMonth[date2]+=element.reward
-        temp[date2]++
-        //gain par mois
-
-        //status des rapports
-        
-        if(_.isUndefined(tempStatus[element.status as keyof typeof tempStatus]))tempStatus[element.status as keyof typeof tempStatus]=0
-        tempStatus[element.status as keyof typeof tempStatus]++
-      })
-        this.statINTI.report_by_month= temp
-        this.statINTI.report_by_status = tempStatus
-        this.statINTI.earn_by_month = tempEarnByMonth
-        this.statINTI.total_rapports=this.scopeINTI.length
-        this.statINTI.average_per_rapport = parseFloat((this.statINTI.earnedEuro /this.scopeINTI.length).toFixed(2))
-        this.loadingINTIGlobal=false
-        //console.log(this.statINTI)
-        this.optionsINTIPie =  this.pieCriticity(this.statINTI.rapport_severity.L,this.statINTI.rapport_severity.M,this.statINTI.rapport_severity.H,this.statINTI.rapport_severity.C)
-
-      
+      this.statINTI = this.globalStat(this.scopeINTI)
+      this.loadingINTIGlobal=false
+      this.optionsINTIPie =  this.pieCriticity(this.statINTI.rapport_severity.L,this.statINTI.rapport_severity.M,this.statINTI.rapport_severity.H,this.statINTI.rapport_severity.C)
       this.initINTIBarre()
       this.initINTIPiReportStatus()
       this.initINTIEarnByMonth()
-      //console.log(this.statINTI)
   }
   syncINTI() {
     this.loadSyncInti=true
@@ -330,60 +279,18 @@ statINTI = {
   }
 
   initINTIPiReportStatus(){
-    let data=Array()
-    Object.keys(this.statINTI.report_by_status).forEach( (element) => {
-      data.push({value:this.statINTI.report_by_status[element as keyof typeof this.statINTI.report_by_status],name:element})
-    })
-    //console.log(data)
+    let data = this.initPie(this.statINTI)
     this.optionsINTIPieReportStatus = this.pieRepport(data)
   }
 
   initINTIBarre(){
-    let month=(new Date()).getMonth();
-    let year=((new Date()).getFullYear()-2);
-    let data=Array()
-    let label=Array()
-    for(var i=0;i<25;i++){
-      let padMonth = `${month}`.padStart(2, '0')
-      let key:string = `${year}${padMonth}`
-      let tempLabel:string = `${year}-${padMonth}`
-      if(Object.prototype.hasOwnProperty.call(this.statINTI.report_by_month,key)){
-       data.push(this.statINTI.report_by_month[key as keyof typeof this.statINTI.report_by_month])
-      } else {
-        data.push(0)
-      }
-      label.push(tempLabel)
-      if(month==12){
-        month=0
-        year++
-      }
-      month++
-    }
-    this.optionsINTIBarre =  this.chartService.barreGraph(data,label,'Report by month','value')
+    let data =this.initBarreReportByMonth(this.statINTI)
+    this.optionsINTIBarre =  this.chartService.barreGraph(data[0],data[1],'Report by month','value')
   }
   
   initINTIEarnByMonth(){
-    let month=(new Date()).getMonth();
-    let year=((new Date()).getFullYear()-2);
-    let data=Array()
-    let label=Array()
-    for(var i=0;i<25;i++){
-      let padMonth = `${month}`.padStart(2, '0')
-      let key:string = `${year}${padMonth}`
-      let tempLabel:string = `${year}-${padMonth}`
-      if(Object.prototype.hasOwnProperty.call(this.statINTI.earn_by_month,key)){
-       data.push(this.statINTI.earn_by_month[key as keyof typeof this.statINTI.earn_by_month])
-      } else {
-        data.push(0)
-      }
-      label.push(tempLabel)
-      if(month==12){
-        month=0
-        year++
-      }
-      month++
-    }
-    this.optionsINTIEarnedByMonth = this.chartService.barreGraph(data,label,"Earn by month","value")
+    let data = this.initBarreEarnedByMonth(this.statINTI)
+    this.optionsINTIEarnedByMonth = this.chartService.barreGraph(data[0],data[1],"Earn by month","value")
   }
   pieRepport(data:any[]){
     return this.chartService.pieChart(data,
