@@ -28,20 +28,11 @@ export class ToolsComponent implements OnInit, AfterViewInit  {
 modalHTML:any;
 
 
-@ViewChild('dialogCreateUserPassword', { read: TemplateRef }) dialogCreateUserPassword:any;
-@ViewChild('dialogCreateApiKey', { read: TemplateRef }) dialogCreateApiKey:any;
-@ViewChild('dialogCreateConfig', { read: TemplateRef }) dialogCreateConfig:any;
-@ViewChild('dialogCreateUserApikey', { read: TemplateRef }) dialogCreateUserApikey:any;
-@ViewChild('dialogCreateUrlApikey', { read: TemplateRef }) dialogCreateUrlApikey:any;
-@ViewChild('dialogCreateWebhook', { read: TemplateRef }) dialogCreateWebhook:any;
+@ViewChild('dialogCreateGlobal', { read: TemplateRef }) dialogCreateGlobal:any;
 private defaultTabButtonsTpl: any;
 
-toolsFormUserPassword: FormGroup = <FormGroup> {};
-toolsFormUserApiKey: FormGroup = <FormGroup> {};
-toolsFormUrlApiKey: FormGroup = <FormGroup> {};
-toolsFormConfig: FormGroup = <FormGroup> {};
-toolsFormApikey: FormGroup = <FormGroup> {};
-toolsFormWebhook: FormGroup = <FormGroup> {};
+
+toolsFormGlobal: FormGroup = <FormGroup> {};
 
   dialogueRefCreate:any;
 
@@ -50,12 +41,7 @@ toolsFormWebhook: FormGroup = <FormGroup> {};
               private toolsService:ToolsService,
               private fbuilder: FormBuilder,
               private dialogService: NbDialogService) {
-    this.toolsFormUserPassword= this.fbuilder.group({ name: '',user: '', password: ''});
-    this.toolsFormApikey= this.fbuilder.group({ name: '',api_key: ''  });
-    this.toolsFormUserApiKey= this.fbuilder.group({ name: '',api_key: '',user:''  });
-    this.toolsFormUrlApiKey= this.fbuilder.group({ name: '',api_key: '',url:''  });
-    this.toolsFormConfig= this.fbuilder.group({ name: '',config_value:''  });
-    this.toolsFormWebhook= this.fbuilder.group({ name: '',webhook:''  });
+    this.toolsFormGlobal= this.fbuilder.group({ name: '',user: '', password: '',api_key:'',webhook:'',url:'',config_value:''});
     this.getToolsModel()
     this.getTools()
 
@@ -73,151 +59,37 @@ toolsFormWebhook: FormGroup = <FormGroup> {};
     this.dialogueRefCreate.close();
   }
   createModal(tool:any){
-
-    let listRequired=Array()
-    this.toolModel[tool].forEach((element:any)=> {
-      listRequired.push(element)
-    })
-    if(listRequired.length==1 && listRequired.includes('config_value')){
-      this.dialogueRefCreate = this.dialogService.open(this.dialogCreateConfig, { context: tool });
-      this.toolsFormConfig.controls['name'].setValue(tool);
-    } 
-    if(listRequired.length==1 && listRequired.includes('api_key')){
-      this.dialogueRefCreate = this.dialogService.open(this.dialogCreateApiKey, { context: tool });
-      this.toolsFormApikey.controls['name'].setValue(tool);
-    } 
-    if(listRequired.length==1 && listRequired.includes('webhook')){
-      this.dialogueRefCreate= this.dialogService.open(this.dialogCreateWebhook, { context: tool });
-      this.toolsFormWebhook.controls['name'].setValue(tool);
-   }
-    if(listRequired.length==2 && listRequired.includes('user') && listRequired.includes('password')){
-      this.dialogueRefCreate= this.dialogService.open(this.dialogCreateUserPassword, { context: tool });
-      this.toolsFormUserPassword.controls['name'].setValue(tool);
-    } 
-    if(listRequired.length==2 && listRequired.includes('api_key') && listRequired.includes('user')){
-       this.dialogueRefCreate= this.dialogService.open(this.dialogCreateUserApikey, { context: tool });
-       this.toolsFormUserApiKey.controls['name'].setValue(tool);
-    }
-    if(listRequired.length==2 && listRequired.includes('api_key') && listRequired.includes('url')){
-      this.dialogueRefCreate= this.dialogService.open(this.dialogCreateUrlApikey, { context: tool });
-      this.toolsFormUrlApiKey.controls['name'].setValue(tool);
-   }
-   
+   console.log(this.toolModel[tool])
+   this.dialogueRefCreate= this.dialogService.open(this.dialogCreateGlobal, { context: {name:tool,required:this.toolModel[tool]} });
+   this.toolsFormGlobal.controls['name'].setValue(tool);
   }
   updateModal(tool:any){
+   this.dialogueRefCreate= this.dialogService.open(this.dialogCreateGlobal, { context: {name:tool,required:this.toolModel[tool]} });
+   this.setElement(tool,this.toolModel[tool])
 
-    let listRequired=Array()
-    this.toolModel[tool].forEach((element:any)=> {
-      listRequired.push(element)
+  }
+  setElement(tool:string,key:any[],){
+    this.tools.forEach((element)=>{ 
+      if(element.name==tool){ 
+        key.forEach((keys:any)=>{
+          if(keys=='config_value'){
+            element.infos[keys as keyof typeof element.infos]=atob(element.infos[keys as keyof typeof element.infos])
+          }
+          this.toolsFormGlobal.controls[keys].setValue(element.infos[keys as keyof typeof element.infos])
+        })
+      } 
     })
-    
-    if(listRequired.length==1 && listRequired.includes('config_value')){
-      this.dialogueRefCreate = this.dialogService.open(this.dialogCreateConfig, { context: tool });
-      var config_value=''
-      this.tools.forEach((element)=>{ 
-        if(element.name==tool){ 
-          config_value= atob(element.infos.config_value)
-        } 
-      })
-      this.toolsFormConfig.controls['config_value'].setValue(config_value);
-      this.toolsFormConfig.controls['name'].setValue(tool);
-    } 
-    if(listRequired.length==1 && listRequired.includes('api_key')){
-      this.dialogueRefCreate = this.dialogService.open(this.dialogCreateApiKey, { context: tool });
-      var api_key=''
-      this.tools.forEach((element)=>{ 
-        if(element.name==tool){ 
-          api_key=element.infos.api_key
-        } 
-      })
-      this.toolsFormApikey.controls['api_key'].setValue(api_key);
-      this.toolsFormApikey.controls['name'].setValue(tool);
-    } 
-    if(listRequired.length==2 && listRequired.includes('webhook') ){
-      this.dialogueRefCreate= this.dialogService.open(this.dialogCreateUrlApikey, { context: tool });
-      var webhook=''
-
-      this.tools.forEach((element)=>{ 
-        if(element.name==tool){ 
-          webhook=element.infos.webhook
-        } 
-      })
-      this.toolsFormUrlApiKey.controls['webhook'].setValue(webhook);
-      this.toolsFormUrlApiKey.controls['name'].setValue(tool);
-   }
-    if(listRequired.length==2 && listRequired.includes('user') && listRequired.includes('password')){
-      this.dialogueRefCreate= this.dialogService.open(this.dialogCreateUserPassword, { context: tool });
-      var user=''
-      var password=''
-      this.tools.forEach((element)=>{ 
-        if(element.name==tool){ 
-          user=element.infos.user
-          password=element.infos.password
-        } 
-      })
-      this.toolsFormUserPassword.controls['user'].setValue(user);
-      this.toolsFormUserPassword.controls['password'].setValue(password);
-      this.toolsFormUserPassword.controls['name'].setValue(tool);
-    } 
-    if(listRequired.length==2 && listRequired.includes('api_key') && listRequired.includes('user')){
-       this.dialogueRefCreate= this.dialogService.open(this.dialogCreateUserApikey, { context: tool });
-       var user=''
-       var api_key=''
-       this.tools.forEach((element)=>{ 
-         if(element.name==tool){ 
-           user=element.infos.user
-           api_key=element.infos.api_key
-         } 
-       })
-       this.toolsFormUserPassword.controls['user'].setValue(user);
-       this.toolsFormUserPassword.controls['password'].setValue(api_key);
-       this.toolsFormUserApiKey.controls['name'].setValue(tool);
-    }
-    if(listRequired.length==2 && listRequired.includes('api_key') && listRequired.includes('url')){
-      this.dialogueRefCreate= this.dialogService.open(this.dialogCreateUrlApikey, { context: tool });
-      var url=''
-      var api_key=''
-      this.tools.forEach((element)=>{ 
-        if(element.name==tool){ 
-          user=element.infos.url
-          api_key=element.infos.api_key
-        } 
-      })
-      this.toolsFormUrlApiKey.controls['url'].setValue(url);
-      this.toolsFormUrlApiKey.controls['password'].setValue(api_key);
-      this.toolsFormUrlApiKey.controls['name'].setValue(tool);
-   }
-   
   }
 
-  inputForm(event:any){
-
-  }
-  updateTools(event:any,type:any){
+  updateTools(event:any,name:any){
     this.loading=true
     event.preventDefault()
-    var data:any={}
-    if(type=='userPassword'){
-       data = this.toolsFormUserPassword.value
-    }
-    else if(type=='apiKey'){
-      data = this.toolsFormApikey.value
-    }
-    else if(type=='config'){
-      data = this.toolsFormConfig.value
+    var data= this.toolsFormGlobal.value
+    if(data.config_value){
       data.config_value=btoa( data.config_value)
     }
-    else if(type=='userApikey'){
-      data = this.toolsFormUserApiKey.value
-    }
-    else if(type=='UrlApikey'){
-      data = this.toolsFormUrlApiKey.value
-    }
-    else if(type=='Webhook'){
-      data = this.toolsFormWebhook.value
-    }
-    let name = data.name
     delete data.name
+    data=Object.fromEntries(Object.entries(data).filter(([_, v]) => v != ""));
     let finalData={"tool":{"name":name,"infos":data}}
     this.toolsService.updateTools(finalData).subscribe( (result) => {
         this.closeModal()
@@ -249,9 +121,7 @@ toolsFormWebhook: FormGroup = <FormGroup> {};
     })
     return data
   }
-  parseToolsRequirement(){
 
-  }
   getToolsModel() {
     this.toolsService.getToolsModel().subscribe( (result) => {
       this.toolModel=result.data
